commit 6

+++ MMul/MMult1x2_3.c
@@ -2,10 +2,10 @@
 #define B(i, j) b[(j) * ldb + (i)]
 #define C(i, j) c[(j) * ldc + (i)]
 
-void AddDot1x2(int, double*, int, double*, int, double*, int);
-void AddDot(int, double*, int, double*, double*);
+void AddDot1x2(int, float*, int, float*, int, float*, int);
+void AddDot(int, float*, int, float*, float*);
 
-void MY_MMult(int m,  int n, int k, double* a, int lda, double* b, int ldb, double* c, int ldc){
+void MY_MMult(int m,  int n, int k, float* a, int lda, float* b, int ldb, float* c, int ldc){
     int i, j;
     for(j = 0; j < n; j += 2){
         for(i = 0; i < m; i++){
@@ -14,14 +14,14 @@ void MY_MMult(int m,  int n, int k, double* a, int lda, double* b, int ldb, doub
     }
 }
 
-void AddDot1x2(int k, double* a, int lda, double* b, int ldb, double* c, int ldc){
+void AddDot1x2(int k, float* a, int lda, float* b, int ldb, float* c, int ldc){
     AddDot(k, &A(0, 0), lda, &B(0, 0), &C(0, 0));
     AddDot(k, &A(0, 0), lda, &B(0, 1), &C(0, 1));
 }
 
 #define X(i) x[(i) * incx]
 
-void AddDot(int k, double* x, int incx, double* y, double* gamma){
+void AddDot(int k, float* x, int incx, float* y, float* gamma){
     int p;
     for(p = 0; p < k; p++){
         *gamma += X(p) * y[p];
